// src/enums/Messages.enum.ts
export enum Message {
  GenericError = "Something wrong happened, please try again later",
  ResourceNotFound = "Resource not found",
  InvalidFileFormat = "Invalid file format",
  InvalidRequestBody = "Invalid request body",
  EmailAlreadyInUse = "Email already in use",
  InvalidPassword = "Password does not match",
  AccessForbidden = "Access forbidden",
  NotAuthenticated = "User not authenticated"
}